Improved the fastapi dependencies classes and added the ability to check permissions for an endpoint.
