+++ frontegg/fastapi/secure_access/__init__.py
@@ -1,5 +1,5 @@
-from .get_user import get_user, GetUser, User
+from .security import FronteggHTTPAuthentication, FronteggSecurity, User
 from .context_provider import context_provider
 from .authentication_middleware import authentication_middleware
 
-__all__ = ('get_user', 'GetUser', 'context_provider', 'authentication_middleware', 'User')
+__all__ = ('FronteggHTTPAuthentication', 'FronteggSecurity', 'context_provider', 'authentication_middleware', 'User')
