+++ test-fast-api.py
@@ -1,5 +1,5 @@
 from frontegg.fastapi import frontegg
-from frontegg.fastapi.secure_access import get_user, User
+from frontegg.fastapi.secure_access import FronteggSecurity, User
 from fastapi import FastAPI, Depends
 import uvicorn
 
@@ -13,7 +13,12 @@ frontegg.init_app(app, client_id=client_id, api_key=api_key, with_secure_access=
 
 
 @app.get("/")
-def read_root(user: User = Depends(get_user)) -> User:
+def read_root(user: User = Depends(FronteggSecurity())) -> User:
+    return user
+
+
+@app.get("/authorized")
+def read_root(user: User = Depends(FronteggSecurity(permissions=['my-permission']))) -> User:
     return user
